dashboard done

+++ app/src/main/java/com/credolabs/justcredo/dashboard/FeedListViewRecyclerAdapter.java
@@ -15,6 +15,7 @@ import com.credolabs.justcredo.model.DbConstants;
 import com.credolabs.justcredo.model.Review;
 import com.credolabs.justcredo.model.User;
 import com.credolabs.justcredo.profile.UserActivity;
+import com.credolabs.justcredo.utility.Util;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.firestore.CollectionReference;
 import com.google.firebase.firestore.DocumentReference;
@@ -52,7 +53,7 @@ public class FeedListViewRecyclerAdapter extends RecyclerView.Adapter<FeedViewHo
         this.arrayList = arrayList;
         this.parent=parent;
 
-        if (parent.equals("blogs")){
+        if (parent.equals(Review.DB_BLOG_REF)){
             likeCollectionRefernce = FirebaseFirestore.getInstance().collection(DbConstants.DB_REF_BLOG_LIKE);
         }else {
             likeCollectionRefernce = FirebaseFirestore.getInstance().collection(DbConstants.DB_REF_REVIEW_LIKE);
@@ -71,6 +72,8 @@ public class FeedListViewRecyclerAdapter extends RecyclerView.Adapter<FeedViewHo
         userCollectionRef.document(uid).get().addOnCompleteListener(userTask -> {
             if (userTask.isSuccessful()){
                 user = userTask.getResult().toObject(User.class);
+            }else{
+                Util.showErrorMessage(context);
             }
         });
     }
@@ -128,7 +131,7 @@ public class FeedListViewRecyclerAdapter extends RecyclerView.Adapter<FeedViewHo
         viewHolder.like.setOnClickListener(v -> {
             likeCollectionRefernce.document(reviewKey).get().addOnCompleteListener(task -> {
                 if (task.isSuccessful()){
-                    if (task.getResult().contains(uid)){
+                    if (task.getResult().exists() && task.getResult().contains(uid)){
                         Map<String, Object> updates = new HashMap<>();
                         updates.put(uid, FieldValue.delete());
                         task.getResult().getReference().update(updates).addOnCompleteListener(task1 -> {
@@ -157,18 +160,12 @@ public class FeedListViewRecyclerAdapter extends RecyclerView.Adapter<FeedViewHo
 
         viewHolder.follow.setOnClickListener(v -> {
             followingCollectionReference.document(uid).get().addOnCompleteListener(task -> {
-                if (task.isSuccessful() && task.getResult().contains(model.getUserID())){
+                if (task.isSuccessful() && task.getResult().exists() && task.getResult().contains(model.getUserID())){
                     AlertDialog.Builder alertDialogBuilder = new AlertDialog.Builder(context);
                     alertDialogBuilder.setMessage("Do you want to unfollow ?");
                     alertDialogBuilder.setPositiveButton("Yes",
                             (arg0, arg1) -> {
-                                final ProgressDialog mProgressDialog = new ProgressDialog(context);
-                                mProgressDialog.setMessage("Processing request...");
-                                mProgressDialog.setProgressStyle(ProgressDialog.STYLE_SPINNER);
-                                mProgressDialog.setIndeterminate(true);
-                                mProgressDialog.setCancelable(false);
-                                mProgressDialog.show();
-
+                                ProgressDialog mProgredialogue = Util.prepareProcessingDialogue(context);
                                 FirebaseFirestore db = FirebaseFirestore.getInstance();
                                 WriteBatch batch = db.batch();
                                 DocumentReference followingRef = followingCollectionReference.document(uid);
@@ -180,7 +177,7 @@ public class FeedListViewRecyclerAdapter extends RecyclerView.Adapter<FeedViewHo
                                 batch.commit().addOnCompleteListener(batchTask -> {
                                     viewHolder.follow.setImageResource(R.drawable.ic_person_add);
                                     FirebaseMessaging.getInstance().unsubscribeFromTopic(model.getUserID());
-                                    mProgressDialog.dismiss();
+                                    Util.removeProcessDialogue(mProgredialogue);
                                 });
                             });
 
@@ -190,13 +187,7 @@ public class FeedListViewRecyclerAdapter extends RecyclerView.Adapter<FeedViewHo
                     alertDialogBuilder.show();
 
                 }else {
-                    final ProgressDialog mProgressDialog = new ProgressDialog(context);
-                    mProgressDialog.setMessage("Processing request...");
-                    mProgressDialog.setProgressStyle(ProgressDialog.STYLE_SPINNER);
-                    mProgressDialog.setIndeterminate(true);
-                    mProgressDialog.setCancelable(false);
-                    mProgressDialog.show();
-
+                    ProgressDialog mProgredialogue = Util.prepareProcessingDialogue(context);
                     FirebaseFirestore db = FirebaseFirestore.getInstance();
                     WriteBatch batch = db.batch();
                     DocumentReference followingRef = followingCollectionReference.document(uid);
@@ -212,7 +203,7 @@ public class FeedListViewRecyclerAdapter extends RecyclerView.Adapter<FeedViewHo
                                 viewHolder.follow.setImageResource(R.drawable.ic_person_black_24dp);
                                 FirebaseMessaging.getInstance().subscribeToTopic(reviewUser.getUid());
                                 User.prepareNotificationFollow(reviewUser,user);
-                                mProgressDialog.dismiss();
+                                Util.removeProcessDialogue(mProgredialogue);
                             });
                         }
                     });
